Added int numBirdsOnBoard() method to Board.java that returns the number of birds on the board
Added int getScore() method to Player.java that returns number of eggs
and number of birds.

In main I made the game stop after 3 rounds.

In main added void printScores() method that prints all the players
scores.

In main in I fixed a bug in drawBirdFromStack method that would cause
the game to crash if all the birds were removed from the stack.